requirements.txt
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
blinker==1.9.0
cachetools==6.2.6
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
deprecation==2.1.0
et_xmlfile==2.0.0
fsspec==2026.3.0
gitdb==4.0.12
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
hyperframe==6.1.0
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mmh3==5.2.1
multidict==6.7.1
mypy_extensions==1.1.0
narwhals==2.18.1
numpy==2.4.4
openpyxl==3.1.5
packaging==26.0
pathspec==1.0.4
patsy==1.0.2
pillow==12.2.0
platformdirs==4.9.4
plotly==6.6.0
pluggy==1.6.0
propcache==0.4.1
protobuf==7.34.1
pyarrow==23.0.1
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.20.0
PyJWT==2.12.1
pyparsing==3.3.2
pyroaring==1.0.4
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytokens==0.4.1
referencing==0.37.0
requests==2.33.1
rich==14.3.3
rpds-py==0.30.0
ruff==0.15.9
scipy==1.17.1
six==1.17.0
smmap==5.0.3
StrEnum==0.4.15
strictyaml==1.7.3
tenacity==9.1.4
toml==0.10.2
tornado==6.5.5
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
watchdog==6.0.0
websockets==15.0.1
xlrd==2.0.2
yarl==1.23.0
zstandard==0.25.0
